+++ elasticapm/traces.py
@@ -32,12 +32,15 @@ import functools
 import logging
 import random
 import re
+import threading
 import time
 import timeit
+from collections import defaultdict
 
 from elasticapm.conf import constants
 from elasticapm.conf.constants import SPAN, TRANSACTION
 from elasticapm.context import init_execution_context
+from elasticapm.metrics.base_metrics import Timer
 from elasticapm.utils import compat, encoding, get_name_from_func
 from elasticapm.utils.disttracing import TraceParent, TracingOptions
 
@@ -55,7 +58,62 @@ TAG_RE = re.compile('[.*"]')
 execution_context = init_execution_context()
 
 
-class Transaction(object):
+class ChildDuration(object):
+    __slots__ = ("obj", "_nesting_level", "_start", "_duration", "_lock")
+
+    def __init__(self, obj):
+        self.obj = obj
+        self._nesting_level = 0
+        self._start = None
+        self._duration = 0
+        self._lock = threading.Lock()
+
+    def start(self, timestamp):
+        with self._lock:
+            self._nesting_level += 1
+            if self._nesting_level == 1:
+                self._start = timestamp
+
+    def stop(self, timestamp):
+        with self._lock:
+            self._nesting_level -= 1
+            if self._nesting_level == 0:
+                self._duration += timestamp - self._start
+
+    @property
+    def duration(self):
+        return self._duration
+
+
+class BaseSpan(object):
+    def __init__(self, tags=None):
+        self._child_durations = ChildDuration(self)
+        self.tags = {}
+        if tags:
+            self.tag(**tags)
+
+    def child_started(self, timestamp):
+        self._child_durations.start(timestamp)
+
+    def child_ended(self, timestamp):
+        self._child_durations.stop(timestamp)
+
+    def end(self, skip_frames=0):
+        raise NotImplementedError()
+
+    def tag(self, **tags):
+        """
+        Tag this span with one or multiple key/value tags. Both the values should be strings
+
+            span_obj.tag(key1="value1", key2="value2")
+
+        Note that keys will be dedotted, replacing dot (.), star (*) and double quote (") with an underscore (_)
+        """
+        for key in tags.keys():
+            self.tags[TAG_RE.sub("_", compat.text_type(key))] = encoding.keyword_field(compat.text_type(tags[key]))
+
+
+class Transaction(BaseSpan):
     def __init__(self, tracer, transaction_type="custom", trace_parent=None, is_sampled=True):
         self.id = "%016x" % random.getrandbits(64)
         self.trace_parent = trace_parent
@@ -64,17 +122,55 @@ class Transaction(object):
         self.duration = None
         self.result = None
         self.transaction_type = transaction_type
-        self._tracer = tracer
+        self.tracer = tracer
 
         self.dropped_spans = 0
         self.context = {}
-        self.tags = {}
 
         self.is_sampled = is_sampled
         self._span_counter = 0
+        self._span_timers = defaultdict(Timer)
+        self._span_timers_lock = threading.Lock()
+        try:
+            self._breakdown = self.tracer._agent._metrics.get_metricset(
+                "elasticapm.metrics.sets.breakdown.BreakdownMetricSet"
+            )
+        except (LookupError, AttributeError):
+            self._breakdown = None
+        try:
+            self._transaction_metrics = self.tracer._agent._metrics.get_metricset(
+                "elasticapm.metrics.sets.transactions.TransactionsMetricSet"
+            )
+        except (LookupError, AttributeError):
+            self._transaction_metrics = None
+        super(Transaction, self).__init__()
 
-    def end_transaction(self):
+    def end(self, skip_frames=0):
         self.duration = _time_func() - self.start_time
+        if self._transaction_metrics:
+            self._transaction_metrics.timer(
+                "transaction.duration",
+                reset_on_collect=True,
+                **{"transaction.name": self.name, "transaction.type": self.transaction_type}
+            ).update(self.duration)
+        if self._breakdown:
+            for (span_type, span_subtype), timer in compat.iteritems(self._span_timers):
+                labels = {
+                    "span.type": span_type,
+                    "transaction.name": self.name,
+                    "transaction.type": self.transaction_type,
+                }
+                if span_subtype:
+                    labels["span.subtype"] = span_subtype
+                self._breakdown.timer("span.self_time", reset_on_collect=True, **labels).update(*timer.val)
+            labels = {"transaction.name": self.name, "transaction.type": self.transaction_type}
+            if self.is_sampled:
+                self._breakdown.counter("transaction.breakdown.count", reset_on_collect=True, **labels).inc()
+                self._breakdown.timer(
+                    "span.self_time",
+                    reset_on_collect=True,
+                    **{"span.type": "app", "transaction.name": self.name, "transaction.type": self.transaction_type}
+                ).update(self.duration - self._child_durations.duration)
 
     def _begin_span(
         self,
@@ -88,7 +184,7 @@ class Transaction(object):
         span_action=None,
     ):
         parent_span = execution_context.get_span()
-        tracer = self._tracer
+        tracer = self.tracer
         if parent_span and parent_span.leaf:
             span = DroppedSpan(parent_span, leaf=True)
         elif tracer.max_spans and self._span_counter > tracer.max_spans - 1:
@@ -103,12 +199,12 @@ class Transaction(object):
                 context=context,
                 leaf=leaf,
                 tags=tags,
+                parent=parent_span,
                 parent_span_id=parent_span_id,
                 span_subtype=span_subtype,
                 span_action=span_action,
             )
             span.frames = tracer.frames_collector_func()
-            span.parent = parent_span
             self._span_counter += 1
         execution_context.set_span(span)
         return span
@@ -137,21 +233,16 @@ class Transaction(object):
         )
 
     def end_span(self, skip_frames=0):
+        """
+        End the currently active span
+        :param skip_frames: numbers of frames to skip in the stack trace
+        :return: the ended span
+        """
         span = execution_context.get_span()
         if span is None:
             raise LookupError()
-        if isinstance(span, DroppedSpan):
-            execution_context.set_span(span.parent)
-            return
 
-        span.duration = _time_func() - span.start_time
-
-        if not self._tracer.span_frames_min_duration or span.duration >= self._tracer.span_frames_min_duration:
-            span.frames = self._tracer.frames_processing_func(span.frames)[skip_frames:]
-        else:
-            span.frames = None
-        execution_context.set_span(span.parent)
-        self._tracer.queue_func(SPAN, span.to_dict())
+        span.end(skip_frames=skip_frames)
         return span
 
     def ensure_parent_id(self):
@@ -164,17 +255,6 @@ class Transaction(object):
             self.trace_parent.span_id = "%016x" % random.getrandbits(64)
             logger.debug("Set parent id to generated %s", self.trace_parent.span_id)
         return self.trace_parent.span_id
-
-    def tag(self, **tags):
-        """
-        Tag this transaction with one or multiple key/value tags. Both the values should be strings
-
-            transaction_obj.tag(key1="value1", key2="value2")
-
-        Note that keys will be dedotted, replacing dot (.), star (*) and double quote (") with an underscore (_)
-        """
-        for key in tags.keys():
-            self.tags[TAG_RE.sub("_", compat.text_type(key))] = encoding.keyword_field(compat.text_type(tags[key]))
 
     def to_dict(self):
         self.context["tags"] = self.tags
@@ -198,8 +278,14 @@ class Transaction(object):
             result["context"] = self.context
         return result
 
+    def track_span_duration(self, span_type, span_subtype, self_duration):
+        # TODO: once asynchronous spans are supported, we should check if the transaction is already finished
+        # TODO: and, if it has, exit without tracking.
+        with self._span_timers_lock:
+            self._span_timers[(span_type, span_subtype)].update(self_duration)
 
-class Span(object):
+
+class Span(BaseSpan):
     __slots__ = (
         "id",
         "transaction",
@@ -216,6 +302,7 @@ class Span(object):
         "parent_span_id",
         "frames",
         "tags",
+        "_child_durations",
     )
 
     def __init__(
@@ -226,6 +313,7 @@ class Span(object):
         context=None,
         leaf=False,
         tags=None,
+        parent=None,
         parent_span_id=None,
         span_subtype=None,
         span_action=None,
@@ -255,12 +343,9 @@ class Span(object):
         # monotonically with respect to the transaction timestamp
         self.timestamp = transaction.timestamp + (self.start_time - transaction.start_time)
         self.duration = None
-        self.parent = None
+        self.parent = parent
         self.parent_span_id = parent_span_id
         self.frames = None
-        self.tags = {}
-        if tags:
-            self.tag(**tags)
         if span_subtype is None and "." in span_type:
             # old style dottet type, let's split it up
             type_bits = span_type.split(".")
@@ -271,17 +356,10 @@ class Span(object):
         self.type = span_type
         self.subtype = span_subtype
         self.action = span_action
-
-    def tag(self, **tags):
-        """
-        Tag this span with one or multiple key/value tags. Both the values should be strings
-
-            span_obj.tag(key1="value1", key2="value2")
-
-        Note that keys will be dedotted, replacing dot (.), star (*) and double quote (") with an underscore (_)
-        """
-        for key in tags.keys():
-            self.tags[TAG_RE.sub("_", compat.text_type(key))] = encoding.keyword_field(compat.text_type(tags[key]))
+        if self.transaction._breakdown:
+            p = self.parent if self.parent else self.transaction
+            p.child_started(self.start_time)
+        super(Span, self).__init__(tags=tags)
 
     def to_dict(self):
         result = {
@@ -307,13 +385,43 @@ class Span(object):
             result["stacktrace"] = self.frames
         return result
 
+    def end(self, skip_frames=0):
+        tracer = self.transaction.tracer
+        timestamp = _time_func()
+        self.duration = timestamp - self.start_time
+        if not tracer.span_frames_min_duration or self.duration >= tracer.span_frames_min_duration:
+            self.frames = tracer.frames_processing_func(self.frames)[skip_frames:]
+        else:
+            self.frames = None
+        execution_context.set_span(self.parent)
+        tracer.queue_func(SPAN, self.to_dict())
+        if self.transaction._breakdown:
+            p = self.parent if self.parent else self.transaction
+            p.child_ended(timestamp)
+            self.transaction.track_span_duration(
+                self.type, self.subtype, self.duration - self._child_durations.duration
+            )
 
-class DroppedSpan(object):
+    def __str__(self):
+        return u"{}/{}/{}".format(self.name, self.type, self.subtype)
+
+
+class DroppedSpan(BaseSpan):
     __slots__ = ("leaf", "parent")
 
     def __init__(self, parent, leaf=False):
         self.parent = parent
         self.leaf = leaf
+        super(DroppedSpan, self).__init__()
+
+    def end(self, skip_frames=0):
+        execution_context.set_span(self.parent)
+
+    def child_started(self, timestamp):
+        pass
+
+    def child_ended(self, timestamp):
+        pass
 
 
 class Tracer(object):
@@ -326,6 +434,7 @@ class Tracer(object):
         max_spans=0,
         span_frames_min_duration=None,
         ignore_patterns=None,
+        agent=None,
     ):
         self.max_spans = max_spans
         self.queue_func = queue_func
@@ -333,6 +442,7 @@ class Tracer(object):
         self.frames_collector_func = frames_collector_func
         self._ignore_patterns = [re.compile(p) for p in ignore_patterns or []]
         self._sample_rate = sample_rate
+        self._agent = agent
         if span_frames_min_duration in (-1, None):
             # both None and -1 mean "no minimum"
             self.span_frames_min_duration = None
@@ -369,9 +479,9 @@ class Tracer(object):
     def end_transaction(self, result=None, transaction_name=None):
         transaction = execution_context.get_transaction(clear=True)
         if transaction:
-            transaction.end_transaction()
             if transaction.name is None:
                 transaction.name = transaction_name if transaction_name is not None else ""
+            transaction.end()
             if self._should_ignore(transaction.name):
                 return
             if transaction.result is None: